Create README.md
If you want to use the code, fork it and change the text_file to whatever you want within the quote marks.

You can also save the audio to an MP3 file. Just go to Convert text to audio. Uncomment the code below, and type or paste your path
where it says "lorem.mp3."